# New S(A,B) indexing