This PR changes the mechanism to match nuclide with S(a,b) table  in the materials. An int-int dictionary of nuclide index to corresponding sab index is created in the material structure. After initialized, this dictionary can tell if a nuclide has a sab table and give the index directly, while the original implementation relies on a sorted list of sab_tables. This new indexing should be more clear and robust, for instance, it will solve the bug of duplicated nuclide with S(a,b) in the issue #676 .