@liangjg 2016-12-23
Well, I implemented the array approach as well. Here are the test results of the three approaches(original, dict, array) with two models, the `tests/test_salphabeta` (4 materials, every material has 1~2 sab tables)  and `beavrs` (16 materials, 2 of them have 1 sab table).

|                 | time/s         | original   | dict       | array      |
|-----------------|----------------|------------|------------|------------|
| test_salphabeta | initialization | 1.7072E+00 | 1.8823E+00 | 1.7060E+00 |
|                 | **transport**      | **2.2278E+03** | **2.3461E+03** | **2.2102E+03** |
|                 | elapsed        | 2.2863E+03 | 2.4297E+03 | 2.3044E+03 |
| | | | | |
| beavrs          | initilaization | 1.2759E+01 | 1.4065E+01 | 1.2402E+01 |
|                 | **transport**      | **4.0304E+03** | **4.0779E+03** | **4.0492E+03** |
|                 | elapsed        | 4.2202E+03 | 4.3052E+03 | 4.2871E+03 |

So it seems the dictionary approach cause 1%~6% performance loss due to the dictionary cost while the array approach has no visible performance loss.