@paulromano 2016-12-30
The array approach you've described could end up wasting a lot of memory for models with many materials with large nuclide inventories, so unfortunately I would have to rule it out.